putting the taxonomy in the edit security wizard to a new page (#117)
